requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
passlib
python-jose
bcrypt
